refactor: 게시물 리스트 더미 데이터 관련 수정
- 뷰모델에서 관리하도록 수정
